Dataset used for the analysis in the paper
"pop" column = the population where the individuals came from. pop1 = Bairro Agronomia; pop2 = Bairro Belém Novo
"species" column = the three species we studied. ATLANTOSCIA FLORIDANA, PETRONIOI and INFLATA
"size.all", "size.fecund", "size.fert" = cephalothorax width of the individuals in mm
"eggs" = number of eggs that the female of that size was carrying;
"embryos" = number of embryos or manca that the female of that size was carrying;
"status" = whether the female was infected with WOLBACHIA or not.